Move exit_actor from worker.py to actor.py
