+++ python/ray/__init__.py
@@ -74,7 +74,6 @@ from ray.worker import (
     connect,
     disconnect,
     error_info,
-    exit_actor,
     get,
     get_gpu_ids,
     get_resource_ids,
@@ -110,7 +109,6 @@ __all__ = [
     "connect",
     "disconnect",
     "error_info",
-    "exit_actor",
     "get",
     "get_gpu_ids",
     "get_resource_ids",
